requirements.txt
requests>=2.32.0
asyncpg
python-telegram-bot
python-dotenv
httpx
aiohttp
requests>=2.32.0

fastapi<0.100
pydantic>=1.10,<2
starlette<0.28
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
asyncpg
httpx
python-dotenv
requests

